+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3]
+
+- Fixed null reference exception when forcing a rule with a tracking callback set and null tracking data.
+- Experiment assignments are included in EvalFeature as well as Run.
+- IGrowthBook interface fixed to inherit IDisposable.
+
 ## [1.0.2]
 
 - Fixed issue with incorrect logic in GrowthBook.GetFeatureResult<T>() call.